# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Casefile Core API",
    "version": "1.0.0",
    "description": "Authoritative internal API contract generated from Zod schemas (D32, D38)."
  },
  "servers": [
    {
      "url": "/v1"
    }
  ],
  "components": {
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "A URI reference that identifies the problem type"
          },
          "title": {
            "type": "string",
            "description": "A short, human-readable summary of the problem type"
          },
          "status": {
            "type": "integer",
            "description": "The HTTP status code generated by the origin server"
          },
          "detail": {
            "type": "string",
            "description": "A human-readable explanation specific to this occurrence of the problem"
          },
          "instance": {
            "type": "string"
          },
          "request_id": {
            "type": "string",
            "description": "Unique identifier for request tracing and correlation"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "request_id"
        ]
      },
...
      "AuthTokenRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "description": "User password"
          },
          "tenantId": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "AuthTokenResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT access token"
          },
          "refreshToken": {
            "type": "string",
            "description": "Rotating refresh token"
          },
          "expiresIn": {
            "type": "number",
            "description": "Access token lifetime in seconds"
          },
          "tokenType": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "tenantId": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "email",
              "name",
              "tenantId"
            ]
          }
        },
        "required": [
          "accessToken",
          "refreshToken",
          "expiresIn",
          "tokenType",
          "user"
        ]
      },
...
    }
  },
  "paths": {
    "/auth/token": {
      "post": {
        "summary": "Exchange credentials for access and refresh tokens",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  }
}